requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
sqlalchemy==2.0.40
mysql-connector-python==9.3.0
pydantic==2.11.4
email_validator==2.2.0
python-jose==3.4.0
passlib==1.7.4
bcrypt==4.3.0
python-multipart==0.0.20
websockets==15.0.1
python-dotenv==1.1.0
cryptography==44.0.3